Refactor Download CSV
